chore: refactor fly
